refactor: splittet ORSGraphManager into several classes in new package
